refactor: move config/parse into global worker
