+++ lib/config/index.spec.ts
@@ -1,7 +1,4 @@
-import upath from 'upath';
 import { getName } from '../../test/util';
-import { readFile } from '../util/fs';
-import getArgv from './config/__fixtures__/argv';
 import { getConfig } from './defaults';
 
 jest.mock('../datasource/npm');
@@ -14,111 +11,6 @@ try {
 const defaultConfig = getConfig();
 
 describe(getName(), () => {
-  describe('.parseConfigs(env, defaultArgv)', () => {
-    let configParser: typeof import('.');
-    let defaultArgv: string[];
-    let defaultEnv: NodeJS.ProcessEnv;
-    beforeEach(async () => {
-      jest.resetModules();
-      configParser = await import('./index');
-      defaultArgv = getArgv();
-      defaultEnv = { RENOVATE_CONFIG_FILE: 'abc' };
-      jest.mock('delay', () => Promise.resolve());
-    });
-    it('supports token in env', async () => {
-      const env: NodeJS.ProcessEnv = { ...defaultEnv, RENOVATE_TOKEN: 'abc' };
-      const parsedConfig = await configParser.parseConfigs(env, defaultArgv);
-      expect(parsedConfig).toContainEntries([['token', 'abc']]);
-    });
-
-    it('supports token in CLI options', async () => {
-      defaultArgv = defaultArgv.concat([
-        '--token=abc',
-        '--pr-footer=custom',
-        '--log-context=abc123',
-      ]);
-      const parsedConfig = await configParser.parseConfigs(
-        defaultEnv,
-        defaultArgv
-      );
-      expect(parsedConfig).toContainEntries([
-        ['token', 'abc'],
-        ['prFooter', 'custom'],
-        ['logContext', 'abc123'],
-      ]);
-    });
-
-    it('supports forceCli', async () => {
-      defaultArgv = defaultArgv.concat(['--force-cli=false']);
-      const env: NodeJS.ProcessEnv = {
-        ...defaultEnv,
-        RENOVATE_TOKEN: 'abc',
-      };
-      const parsedConfig = await configParser.parseConfigs(env, defaultArgv);
-      expect(parsedConfig).toContainEntries([
-        ['token', 'abc'],
-        ['force', null],
-      ]);
-      expect(parsedConfig).not.toContainKey('configFile');
-    });
-    it('supports config.force', async () => {
-      const configPath = upath.join(
-        __dirname,
-        'config/__fixtures__/with-force.js'
-      );
-      const env: NodeJS.ProcessEnv = {
-        ...defaultEnv,
-        RENOVATE_CONFIG_FILE: configPath,
-      };
-      const parsedConfig = await configParser.parseConfigs(env, defaultArgv);
-      expect(parsedConfig).toContainEntries([
-        ['token', 'abcdefg'],
-        [
-          'force',
-          {
-            schedule: null,
-          },
-        ],
-      ]);
-    });
-    it('reads private key from file', async () => {
-      const privateKeyPath = upath.join(
-        __dirname,
-        'keys/__fixtures__/private.pem'
-      );
-      const env: NodeJS.ProcessEnv = {
-        ...defaultEnv,
-        RENOVATE_PRIVATE_KEY_PATH: privateKeyPath,
-      };
-      const expected = await readFile(privateKeyPath);
-      const parsedConfig = await configParser.parseConfigs(env, defaultArgv);
-
-      expect(parsedConfig).toContainEntries([['privateKey', expected]]);
-    });
-    it('supports Bitbucket username/passwod', async () => {
-      defaultArgv = defaultArgv.concat([
-        '--platform=bitbucket',
-        '--username=user',
-        '--password=pass',
-      ]);
-      const parsedConfig = await configParser.parseConfigs(
-        defaultEnv,
-        defaultArgv
-      );
-      expect(parsedConfig).toContainEntries([
-        ['platform', 'bitbucket'],
-        ['username', 'user'],
-        ['password', 'pass'],
-      ]);
-    });
-    it('massages trailing slash into endpoint', async () => {
-      defaultArgv = defaultArgv.concat([
-        '--endpoint=https://github.renovatebot.com/api/v3',
-      ]);
-      const parsed = await configParser.parseConfigs(defaultEnv, defaultArgv);
-      expect(parsed.endpoint).toEqual('https://github.renovatebot.com/api/v3/');
-    });
-  });
   describe('mergeChildConfig(parentConfig, childConfig)', () => {
     it('merges', async () => {
       const parentConfig = { ...defaultConfig };
