+++ lib/workers/global/config/parse/__fixtures__/file.js
