+++ lib/workers/global/config/parse/__fixtures__/with-force.js
